+++ src/sync/cache.rs
@@ -1,8 +1,7 @@
-use super::ConcurrentCacheExt;
-use crate::common::{
+use super::{
     base_cache::{BaseCache, HouseKeeperArc, MAX_SYNC_REPEATS, WRITE_RETRY_INTERVAL_MICROS},
     housekeeper::InnerSync,
-    WriteOp,
+    ConcurrentCacheExt, WriteOp,
 };
 
 use crossbeam_channel::{Sender, TrySendError};
@@ -425,12 +424,7 @@ mod tests {
     #[test]
     fn basic_multi_threads() {
         let num_threads = 4;
-
-        let mut cache = Cache::new(100);
-        cache.reconfigure_for_testing();
-
-        // Make the cache exterior immutable.
-        let cache = cache;
+        let cache = Cache::new(100);
 
         let handles = (0..num_threads)
             .map(|id| {
@@ -438,7 +432,6 @@ mod tests {
                 std::thread::spawn(move || {
                     cache.insert(10, format!("{}-100", id));
                     cache.get(&10);
-                    cache.sync();
                     cache.insert(20, format!("{}-200", id));
                     cache.invalidate(&10);
                 })
@@ -446,8 +439,6 @@ mod tests {
             .collect::<Vec<_>>();
 
         handles.into_iter().for_each(|h| h.join().expect("Failed"));
-
-        cache.sync();
 
         assert!(cache.get(&10).is_none());
         assert!(cache.get(&20).is_some());